ssh: add functionality to establish ssh tunnels
This sub-command can be used to request ssh connections to possibly NATed
devices. To this end, this command uses the oauth2 authentication mechanism
to request an access token and then requests the ssh tunnel from the
backend. As a result the omnect-cli will produce two temporary
certificates for the connection, as well as a ssh configuration file in
a temporary directory, which the user can then use to connect to the
device.