release 0.16.0: add ssh command (#80)
* tests: clippy error

* tests: cargo fmt

* cargo.toml: remove newline

* auth: add functionality to perform oauth2 authentication

Some CLI functionalities will require Keycloak access tokens for their
operation. This adds functionality to authenticate against the Keycloak backend
to retrieve such an access token. The implemented authentication mechanism
thereby is the oauth2 authorization code flow.

The authentication mechanism furthermore attempts to store the refresh token in
the system key ring (if available) to evade future authentication requests.

* ssh: add functionality to establish ssh tunnels

This sub-command can be used to request ssh connections to possibly NATed
devices. To this end, this command uses the oauth2 authentication mechanism
to request an access token and then requests the ssh tunnel from the
backend. As a result the omnect-cli will produce two temporary
certificates for the connection, as well as a ssh configuration file in
a temporary directory, which the user can then use to connect to the
device.

* ssh: add test cases for ssh command

* ssh: add instructions about ssh feature to readme

* ssh: add validators for ssh config

* ssh: add command to inject ssh CAs

* readme: add ssh cert injection information

* auth: change from hyper to actix

* release 0.16.0