feat(ssh): enable ssh tunnel from containerized environments
When running the omnect-cli from within a containerized environment,
authentication will not work out of the box: when binding to localhost, the
client will bind to localhost from inside of the container, it is not possible
to connect from the container externally.

This adds a check for the "CONTAINERIZED" environment variable. If this
environment variable is set, the auth redirect server will instead bind to
"0.0.0.0", the idea being, that docker can do the mapping to local host, then.

Furthermore, we add the "CONTAINER_HOST" variable which allows the omnect-cli to
determine the host system so that an according ssh config can be generated.