feat(ssh): add ssh tunnel support for the docker image (#114)
When running omnect-cli in a docker container, we need some extra steps
to get the ssh tunnel feature to work. Furthermore, when running on
windows host systems we have to make some adjustments in the ssh
configuration.

This PR extends the omnect-cli so that we can pass the setup
(containerized or not, windows host) in via environment variables and to
generate an according configuration.

Furthermore, this extends the documentation to include the necessary
steps to run the omnect-cli in a containerized environment.