feat(ssh): query user to overwrite config
With this we query the user whether they want to overwrite their pre-existing
configuration file. This resolves the issue of omnect-cli getting into the way
of the user, while still providing the safety net of not overwriting
existing (user)-configs with temporarily valid generated configs.

Furthermore, we now only perform the check if the user provided a custom config
path. This is under the assumption, that if they want to use the default path,
they don't care whether ssh configs there are overwritten.