feat(ssh_tunnel): add support for authentication via ipv4 and ipv6 (#121)
Some users of the situation faced the situation where redirect link the
oauth2 flow was not invoked correctly. The problem here was that
localhost was resolved differently, depending on the respective user's
machine (either as IPv4 only or as IPv6 only). If the redirect server,
however, bound to the opposing protocol, the resolver link did not work.

To the best of our knowledge, this is an issue with actix (or underlying
crates), as the "bind" method promises to bind to either IPv4 and IPv6
if "localhost" is specified.

With this fix we resolve the issue by extending the authentication info
so that multiple bind addresses can be specified. We then explicitly
specify both, the IPv4 and the IPv6 localhost addresses.