feat(cli,ecos): add set-ecos-config cli command stub
We add a CLI command stub for the ECOS certificate use case. When finished, the
command should automatically create a key pair, create a CSR for the key pair,
retrieve the certificate for this CSR from the ECOS server and inject the data
into the firmware image.

The command, therefore, has the typical image options (image name, packing the
image, creation of a bmap file), user credentials for the ECOS service, and the
identifier for the respective ECOS ca.