refactor(file): make set_device_cert generic over certificate data
set_device_cert requires certificate/key data from three different input
parameters. Currently, the first parameter is expected to be a file, whereas the
latter two are expected to be u8 vectors. We introduce a new trait "CertData"
and implement it for paths and vectors in order to homogenize the interface.