doc(ssh_tunnel): fix config path for docker setup
We changed the base directory for generated ssh configuration files to be
consistent in the code base. With this change, the path of the configuration
file in the docker example changed, as well. The documentation currently still
shows an example with the old path, thus causing the example to not work. We
update the example so that it is functional again.

Signed-off-by: Tobias Langer <tobias.langer@conplement.de>