fix(auth): resolve freeze on errors in redirect server
The omnect-cli would freeze when encountering errors, such as already bound
ports and alike, when setting up the redirect server. The issue was that we
would first wait for the server to be up, before then checking errors. We
resolved the issue by extending the "server-up" message with a failure message
type.