chore: add ignore for RUSTSEC-2024-0370
proc-macro-error is unmaintained of now. The authors of the validator
crate are aware of the issue and working on it
(https://github.com/Keats/validator/issues/351) as of now, however, there
is no solution. As this issue is not critical for operation we ignore it
for now and update the affected dependencies when the updates are
available.