feat(ssh): add support for multiple ssh connections
Until now the omnect-cli would overwrite ssh configs, keys, and certificates.
This had the consequence, that one could only ever use a single configuration at
a time (if they wouldn't explicitly specify the file paths to use). We now
prefix these file paths with the device names in order to lift this limitation.