fix(ssh): failure create https connection to keycloak (#149)
The omnect-cli is not able to create ssh connections, as it cannot
create secure connections to the keycloak server to retrieve an access
token. This regression was introduced when default features were
disabled for all dependencies. One consequence of this action was, that
reqwest did no longer ship with tls support. We therefore re-enable said
feature.